chore (scaffold) update android project
