+++ android/app/src/main/java/com/opencity/MainApplication.java
@@ -1,0 +1,55 @@
+package com.opencity;
+
+import android.app.Application;
+
+import com.facebook.react.ReactApplication;
+import io.realm.react.RealmReactPackage;
+import com.airbnb.android.react.maps.MapsPackage;
+import com.babisoft.ReactNativeLocalization.ReactNativeLocalizationPackage;
+import fr.bamlab.rnimageresizer.ImageResizerPackage;
+import com.imagepicker.ImagePickerPackage;
+import com.facebook.react.ReactNativeHost;
+import com.facebook.react.ReactPackage;
+import com.facebook.react.shell.MainReactPackage;
+import com.facebook.soloader.SoLoader;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class MainApplication extends Application implements ReactApplication {
+
+  private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
+    @Override
+    public boolean getUseDeveloperSupport() {
+      return BuildConfig.DEBUG;
+    }
+
+    @Override
+    protected List<ReactPackage> getPackages() {
+      return Arrays.<ReactPackage>asList(
+          new MainReactPackage(),
+            new RealmReactPackage(),
+            new MapsPackage(),
+            new ReactNativeLocalizationPackage(),
+            new ImageResizerPackage(),
+            new ImagePickerPackage()
+      );
+    }
+
+    @Override
+    protected String getJSMainModuleName() {
+      return "index";
+    }
+  };
+
+  @Override
+  public ReactNativeHost getReactNativeHost() {
+    return mReactNativeHost;
+  }
+
+  @Override
+  public void onCreate() {
+    super.onCreate();
+    SoLoader.init(this, /* native exopackage */ false);
+  }
+}
