--- android/app/src/main/res/values/api-keys.example.xml
@@ -1,4 +1,0 @@
-<resources>
-    <!--<string name="GoogleMapsKey">InsertGoogleMapsAndroidAPIKeyHere</string>-->
-</resources>
-<!-- Copy/rename this file to api-keys.xml, remove comments -->
