--- android/app/src/main/res/values/colors.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <color name="primaryColorDark">#005eb8</color>
-    <color name="primaryColorLight">#fafafa</color>
-
-</resources>
