+++ android/app/src/main/res/values/styles.xml
@@ -3,7 +3,6 @@
     <!-- Base application theme. -->
     <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
         <!-- Customize your theme here. -->
-        <item name="colorPrimaryDark">@color/primaryColorLight</item>
-        <item name="android:windowLightStatusBar">true</item>
     </style>
+
 </resources>
